Move metrics tenant isolation to metrics v1 pkg
